Add files via upload
We present a cellular automaton model that efficiently tracks a target in distributed mobile wireless sensor network using Hexagonal Grid. 
We focus on simulating the mobile wireless sensor networks for tracking a moving target considering energy-efficiency. Using the mobility
of sensors, we first disperse sensors as far as possible for better coverage while preserving connectivity. We use Motion Planning Algorithm for Sensor dispersing. 
Furthermore, we introduce a trilevel alert system for activating sensors selectively. We conduct experiments for evaluating the performance of the proposed model. The performance of our model is measured in terms of lifetime, coverage, and tracking quality, and compared with other models.